+++ rust/src/core/mod.rs
@@ -281,6 +281,7 @@ where
         })
     }
 
+    #[allow(dead_code)]
     pub(crate) fn with_map<MI2: Metric, MO2: Measure>(
         &self,
         input_metric: MI2,
