+++ rust/src/traits/mod.rs
@@ -14,6 +14,9 @@ pub use arithmetic::*;
 mod cast;
 pub use cast::*;
 
+mod domains;
+pub use domains::*;
+
 mod operations;
 pub use operations::*;
